requirements.txt
gradio==4.44.0
google-generativeai==0.7.2
python-dotenv==1.0.1
pypdf==5.0.0
pandas
